consul-template 0.18.3
Updates consul-template to 0.18.3.

The service file has been removed, because it's expected that child
images will create their own services which invoke run-consul-template,
specifying a config file that uses consul-template's "exec" mode.

The service file was convenient for images that only run a single
process managed by consul-template, but a lot of our images don't, so
they'd have to stop or remove this service file, which is cumbersome.
Also, it's better that the default service name isn't "consul-template,"
but named according to the actual service it manages.

The file consul_template_helpers.sh was removed because the
wait_for_file function is useless with exec mode.