Dedup support
Allow downstream images to supply a script that generates the dedup
prefix. This is done instead of reading an environment variable, because
e.g. in the case of MBS, the dedup prefix must be a combination of the
DEPLOY_ENV and GIT_SHA, which can't be simultaneously known outside the
running container.